add flask error support
